Move rendering related things to PlatformContext
