+++ inventory-framework-api/src/main/java/me/devnatan/inventoryframework/UpdateReason.java
@@ -1,0 +1,3 @@
+package me.devnatan.inventoryframework;
+
+public interface UpdateReason {}
